Cleaned up pages by moving functions into separate files to be reused, added page for gloss instance, fixed a couple bugs
