+++ package.json
@@ -25,7 +25,7 @@
     "eslint": "8.50.0",
     "eslint-config-next": "13.5.3",
     "lucide-react": "^0.279.0",
-    "next": "13.5.3",
+    "next": "^13.5.5",
     "next-themes": "^0.2.1",
     "postcss": "8.4.30",
     "react": "18.2.0",
@@ -34,6 +34,7 @@
     "tailwind-merge": "^1.14.0",
     "tailwindcss": "3.3.3",
     "tailwindcss-animate": "^1.0.7",
-    "typescript": "5.2.2"
+    "typescript": "5.2.2",
+    "zod": "^3.22.4"
   }
 }
