+++ src/configs/rerum-links.ts
@@ -1,2 +1,4 @@
 export const PRODUCTION_GLOSS_COLLECTION =
   "https://store.rerum.io/v1/id/610c54deffce846a83e70625";
+
+export const RERUM = "https://store.rerum.io/v1/id/";
