+++ src/lib/validators/TargetId.ts
@@ -1,0 +1,5 @@
+import { z } from "zod";
+
+export const TargetIdValidator = z.object({
+  targetId: z.string(),
+});
